requirements.txt
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
Pillow==9.4.0
Pillow==10.0.0
rembg==2.0.50
scipy==1.10.1
